% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sarm.R
\name{predict_sarm}
\alias{predict_sarm}
\title{Get Predicted Values from a SARM Model}
\usage{
predict_sarm(model, newdata = NULL)
}
\description{
This function reports predicted values with 95% confidence intervals
for a model estimated with the sarm() function. By default, it
generates predicted values with the data used to estimate the model.
The user may create or use a different dataset, and then include
this using the newdata command.
}
