% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rob_sarm.R
\name{rob_sarm}
\alias{rob_sarm}
\title{Estimate a Strategic Autoregressive Model with Modeled Variance}
\usage{
rob_sarm(eq, data = NULL)
}
\description{
This function is an extension of the sarm() model. It allows the user
to estimate the sarm() model but with model variance given as a function
of a grouping variable. This approach may be adopted if the user
suspects the data violate the assumptions of constant variance and
independence of observations within groups. When estimating this model,
it is assumed that the last variable included on the right-hand side
of the equation object is the grouping variable.
}
