% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarm.R
\name{sarm}
\alias{sarm}
\title{Estimate a Strategic Autoregressive Model}
\usage{
sarm(eq, data = NULL, robust = F, cluster = F)
}
\description{
This function takes a user supplied equation object and
dataframe. It assumes that the first two variables on the
right-hand side of the equation are the actor-specific
resource constraint and the spatial lag (sum of other-actor
expenditures on the strategic good), respectively. The
remaining variables are assumed to predict intrinsic
preference for the strategic good. NA values are allowable.
By default, it is assumed that estimated standard errors are not
robust, nor are they clustered according to some grouping variable.
If robust standard errors are desired, the user should set
robust = TRUE. If robust-clustered standard errors are desired,
the user should set both robust = TRUE and cluster = TRUE. If
clustered standard errors are desired, it is assumed that the last
variable included on the right-hand side of the equation is the
grouping variable. The grouping variable may be numeric, a factor,
or a character string.
}
