% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sarm.R
\name{summary_sarm}
\alias{summary_sarm}
\title{Report Model Summary Statistics for SARM}
\usage{
summary_sarm(model)
}
\description{
This function returns summary output from a model estimated with
sarm(). It acts much like the summary.lm() function.
}
